for python file you can reach out to me
India in 2023 a 3 Trillion $ economy became worlds fastest growing economy with global ranking 5 and by 2027 aiming for rank 3. 



Having said that for any country's most important factors to grown are Increase export and reduce Trade deficit. 



here i have data of 2010-2018 of Indian trade and have some very interesting  finding. how we are doing in terms of Exports and deficits and with which country we have highest trades and which commodities are we exporting the most



Findings-

- We crossed mark of $350 Billion export value in 2018

- Highest export is to USA $350 B, followed by UAE $ 290B from 2010-2018

- While Minerals, oils are the highest exported commodity approx $ 410 B, cotton is least $ 50 B.

- Import for year 2018 alone crossed $ 450 B

- in 2018 trade deficit was approx. $ 180 B. Import stand at $500 B and export $320 B.

- Trade surplus countries are USA, Bangladesh, Netherlands, Nepal & Sri Lanka

- Trade deficit countries are China, Switzerland, Saudi, Iraq & Qatar

- Top 5 trade partners China, USA, UAE, Saudi and Switzerland

- With 59.5% countries India's trade have been surplus and with 40.5% countries India have Trade deficit